cp-170: + usePressState hook
