+++ mobile/src/libs/hooks/hooks.ts
@@ -4,6 +4,7 @@ export { useAppRoute } from './use-app-route/use-app-route.hook';
 export { useAppSelector } from './use-app-selector/use-app-selector.hook';
 export { usePlayerControls } from './use-player-controls/use-player-controls.hook';
 export { usePlayerProgress } from './use-player-progress/use-player-progress.hook';
+export { usePressState } from './use-press-state/use-press-state.hook';
 export { useSearch } from './use-search/use-search.hook';
 export { useFocusEffect, useNavigation } from '@react-navigation/native';
 export { useCallback, useEffect, useMemo, useRef, useState } from 'react';
